make token balance fetch reuseable
