+++ packages/commonwealth/client/scripts/controllers/app/login.ts
@@ -16,6 +16,7 @@ import type ChainInfo from '../../models/ChainInfo';
 import SocialAccount from '../../models/SocialAccount';
 import { CosmosExtension } from '@magic-ext/cosmos';
 import BN from 'bn.js';
+import { getTokenBalance } from 'client/scripts/helpers/token_balance_helper';
 
 export function linkExistingAddressToChainOrCommunity(
   address: string,
@@ -182,33 +183,8 @@ export async function updateActiveAddresses({
       .filter((addr) => addr),
     shouldRedraw
   );
-  if (
-    app.user.activeAccounts[0] &&
-    app.chain.gatedTopics?.length > 0 &&
-    !app.user.activeAccounts[0].tokenBalance
-  ) {
-    try {
-      const contract = app.contracts.getCommunityContracts();
-      $.post(`${app.serverUrl()}/tokenBalance`, {
-        chain: app.chain.meta.id,
-        address: app.user.activeAccounts[0].address,
-        author_chain: app.chain.meta.id,
-        contract_address: contract.length > 0 ? contract[0].address : null,
-        all: true,
-      }).then((balanceResp) => {
-        if (balanceResp.result) {
-          balanceResp.result.forEach((balObj) => {
-            const account = app.user.activeAccounts.find(
-              (acc) => acc.address == balObj.address.distinctAddress
-            );
-            if (account) account.setTokenBalance(new BN(balObj.balance, 10));
-          });
-        }
-      });
-    } catch {
-      console.log('address not compatible');
-    }
-  }
+
+  getTokenBalance();
 
   // select the address that the new chain should be initialized with
   const memberAddresses = app.user.activeAccounts.filter((account) => {
